Rename file to entry.rb
* Remove .rb from requires
* Move requires to top of file
