+++ lib/rgeo/coord_sys/srs_database/entry.rb
